Factor State\Data out of ResultData.
